@@ -7,11 +7,11 @@ classifiers =
     Intended Audience :: Developers,
     Natural Language :: English,
     Programming Language :: Python :: 3,
-    Programming Language :: Python :: 3.8,
+    Programming Language :: Python :: 3.10,
 
 
 [options]
 packages = find:
-python_requires = >=3.8
+python_requires = >=3.10
 setup_requires =
     setuptools_scm
