@@ -57,7 +57,7 @@ def test_holo_pipeline():
                                                                          pyopia.instrument.holo.load_image),
                  'load': holo.Load(),
                  'correct background': pyopia.background.CorrectBackgroundAccurate(pyopia.background.shift_bgstack_accurate),
-                 'reconstruct': holo.Reconstruct(stack_clean=0.02),
+                 'reconstruct': holo.Reconstruct(stack_clean=0.02, forward_filter_option=2, inverse_output_option=0),
                  'focus': holo.Focus(pyopia.instrument.holo.std_map,
                                      threshold=threshold,
                                      focus_function=pyopia.instrument.holo.find_focus_sobel,
@@ -83,7 +83,7 @@ def test_holo_pipeline():
         print('stats header: ', stats.columns)
         print('Total number of particles: ', len(stats))
         assert len(stats) == 42, ('Number of particles expected in this test is 42.' +
-                                  'This test counted' + len(stats) +
+                                  'This test counted' + str(len(stats)) +
                                   'Something has altered the number of particles detected')
 
 
@@ -137,7 +137,7 @@ def test_silcam_pipeline():
         assert num_images == 1, ('Number of images expected is 1.' +
                                  'This test counted' + str(num_images))
         assert len(stats) == 870, ('Number of particles expected in this test is 870.' +
-                                   'This test counted' + len(stats) +
+                                   'This test counted' + str(len(stats)) +
                                    'Something has altered the number of particles detected')
 
 
